dZEndpoint should be a file-wide constant
